Fixes from eclipse build. Problems with alib.str.*Escape
Mostly these are fixes to warnings from eclipse. Usually this is unused
vars, unused imports or issues with generics.

As eclipse was configured to build with 1.5 which we claim to support
there are some issues that are specific to 1.5 vs. 1.6. These include
File#canExecute and @Override application to interface implementations.

There are issues that need fixing in the Escape stuff in alib.str. First
and foremost we should combine our stuff from esapi into a better
framework for escaping in general. Secondly I've renamed XEscape#XEscape
to XEscape#escape.